Add damage care
To get rid of damage from highly concentrated gas.

EtOH is very common gas. normally, organic solvent is not permitted to clean the surface of product. but accidentally, when user use that solvent to clean something nearby the sensor, sensor can be damaged to their raw signal of sensing data of the carbon dioxide gas.

so prevent changing the raw signal from damage and add recovery sequence to code. 

2020.01.13